requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
requests
python-multipart
python-dotenv
